+++ docs/content/en/changelog.md
@@ -10,17 +10,19 @@ All notable changes to this project will be documented in this file. For a full 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-### Unreleased
+### [v1.5.0](https://github.com/pestphp/defstudio-plugin-laravel-expectations/compare/v1.4.0...v1.5.0) - 2021-10-25
 
 **Added**
 
 - `toBeAfter()` expectation
 - `toBeBefore()` expectation
 - `toBeBetween()` expectation
+- `toBeCurrentDay()` expectation
 - `toBeCurrentMonth()` expectation
 - `toBeCurrentWeek()` expectation
 - `toBeCurrentYear()` expectation
 - `toBeFriday()` expectation
+- `toBeFuture()` expectation
 - `toBeFuture()` expectation
 - `toBeLastMonth()` expectation
 - `toBeLastWeek()` expectation
@@ -37,13 +39,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `toBeSameWeekAs()` expectation
 - `toBeSaturday()` expectation
 - `toBeSunday()` expectation
+- `toBeToday()` expectation
 - `toBeThursday()` expectation
 - `toBeTuesday()` expectation
 - `toBeWednesday()` expectation
 - `toBeWeekday()` expectation
 - `toBeWeekend()` expectation
+- `toHaveAllSession()` expectation
 - `toHaveHeader()` expectation
 - `toHaveSession()` expectation
+- `toBeMissingInStorage()` expectation
 
 **Updated**